backend/requirements.txt
Flask==3.1.1
flask_sock==0.7.0
flask_sqlalchemy==3.1.1
google_api_python_client==2.169.0
google_auth_oauthlib==1.2.1
langchain==0.3.25
livekit==1.0.6
livekit_api==1.0.2
livekit_protocol==1.0.2
openai==1.79.0
protobuf==6.31.0
python-dotenv==1.1.0
qdrant_client==1.14.2
Requests==2.32.3
speechrecognition==3.14.2
stripe==11.3.0
twilio==9.6.0
websockets==15.0.1
Werkzeug==3.1.3
gunicorn
gevent
psycopg2
flask-cors
alembic